Ensure context is made current before WM_DESTROY
The WM_DESTROY message is sent to the window procedure of the window
being destroyed after the window is removed from the screen.

In certain drivers, making the current GL context current is still
possible during WM_DESTROY handling but in other drivers it can crash
the process.

Since we need the context to be current during shutdown, we move setting
the context to WM_CLOSE. However, note this message is not sent to child
controls, but only to the parent form, so we have to intercept it there.