Allow adjusting amplitude range limits
By default the visualizer auto-fits axis limits to the data on a
per-channel basis. Here we allow specifying an amplitude range limit,
and an optional range offset. This allows having a single slider to
adjust scale similar to an oscilloscope.

Optional label property is available to label units of measure for
convenience.